blocks/wavfile_source: fix case where we'll loop producing zero forever
This required checking for actual return amount in-loop; also, do not assume that length of a file fits in an int; there's files longer than 2³¹.

Be clear about ignored sndfile errors.

+++ gr-blocks/include/gnuradio/blocks/wavfile.h
@@ -29,10 +29,10 @@ struct wav_header_info {
     int nchans;
 
     //! Bytes per sample
-    int bytes_per_sample;
+    size_t bytes_per_sample;
 
     //! Number of samples per channel
-    long long samples_per_chan;
+    size_t samples_per_chan;
 
     //! sndfile format
     int format;
